created reusable hook for handling api call
